moved emitDts functionality into svelte2tsx, adjust accordingly
