+++ packages/kit/package.json
@@ -33,7 +33,7 @@
 		"selfsigned": "^1.10.11",
 		"sirv": "^1.0.12",
 		"svelte": "^3.38.2",
-		"svelte2tsx": "~0.4.0",
+		"svelte2tsx": "~0.4.1",
 		"svelte-check": "^2.2.0",
 		"tiny-glob": "^0.2.8",
 		"typescript": "^4.2.4",
